eql/3

+++ src/lib.c
@@ -3,7 +3,7 @@
 // freestanding libc substitutes
 
 intptr_t lcprng(intptr_t s) {
-  const intptr_t steele_vigna_2021 = 0xaf251af3b0f025b5;
+  const int64_t steele_vigna_2021 = 0xaf251af3b0f025b5;
   return (s * steele_vigna_2021 + 1) >> 8; }
 
 void setw(void *x, uintptr_t i, size_t l) {
